debug onet

+++ dface/core/detect.py
@@ -524,7 +524,7 @@ face candidates:%d, current batch_size:%d"%(num_boxes, batch_size)
         print('feed_imgs = feed_imgs.cuda()')
         if self.rnet_detector.use_cuda:
             print('self.rnet_detector.use_cuda')
-            feed_imgs = feed_imgs.cuda()
+            feed_imgs = feed_imgs.float().cuda()
         print('before onet detector')
         cls_map, reg, landmark = self.onet_detector(feed_imgs)
         print('after onet detector')
